Move search code around to break circular import
search.py was getting too large and needed a refactor prior to some
development work. The `Wordform.*search*` static methods didn’t reference
*anything* in the `Wordform` class, and having them caused all sorts of
circular import trouble when the `Wordform` would import search, but search
would import `APIConfig` to try to preload a cache, and `APIConfig` would
need `models.Wordform` to be loaded first.
