+++ CreeDictionary/API/views.py
@@ -1,8 +1,9 @@
 from typing import List
 
-from API.schema import SerializedSearchResult
 from django.http import HttpResponse, HttpResponseBadRequest, JsonResponse
-from .models import Wordform
+
+from API.schema import SerializedSearchResult
+from .search import simple_search
 
 
 def click_in_text(request) -> HttpResponse:
@@ -18,7 +19,7 @@ def click_in_text(request) -> HttpResponse:
         return HttpResponseBadRequest("query param q is an empty string")
 
     results: List[SerializedSearchResult] = []
-    for result in Wordform.simple_search(q, include_auto_definitions=False):
+    for result in simple_search(q, include_auto_definitions=False):
         results.append(result.serialize(include_auto_definitions=False))
 
     response = {"results": results}
